chart dynamically graphs path of station
