+++ src/components/compass/index.js
@@ -1,0 +1,3 @@
+import Compass from './Compass';
+
+export default Compass
