+++ src/components/compass/style.scss
@@ -1,0 +1,4 @@
+svg {
+    // width: 145px;
+    // padding: 0px 25px 0px 25px;
+}
