+++ src/components/sightingtable/style.scss
@@ -81,19 +81,15 @@
                     margin-top: 9px;
                 }
             }
-            span {
-                // margin-left: 1%;
-                img {
-                    width: 145px;
-                    padding: 0px 25px 0px 25px;
-                    border-radius: 29%;
-                }
-            }
+            // span {
+            //     svg {
+            //         width: 145px;
+            //         padding: 0px 25px 0px 25px;
+            //         border-radius: 29%;
+            //     }
+            // }
         }
 
-        // #sightingdownarrow {
-        //     transform: scale(.5);
-        // }
 
         #headerspacingspan {
             width: 10%;
